- added page title to GUI_Panel_Main - fixed: trying to delete non-existant files doesn't result in exception

+++ io/file.php
@@ -40,6 +40,8 @@ class IO_File {
 	}
 	
 	public function delete() {
+		if (!$this->exists())
+			return true;
 		return unlink($this->file);
 	}
